Add TR2 PSX Support Load TR2 PSX levels. Also move each level loader into its own file to make things clearer.
