+++ trlevel/tr_rooms.cpp
@@ -92,6 +92,26 @@ namespace trlevel
         return new_vertices;
     }
 
+    std::vector<trview_room_vertex> convert_vertices_tr2_psx(std::vector<uint32_t> vertices, int32_t y_top)
+    {
+        return vertices |
+            std::views::transform([=](auto&& v)
+                {
+                    const float colour = static_cast<float>(((v >> 15) & 0x7fff)) / 0x7fff;
+                    return trview_room_vertex
+                    {
+                        .vertex =
+                        {
+                            .x = static_cast<int16_t>(((v >> 10) & 0x1f) << 10),
+                            .y = static_cast<int16_t>((((v >> 5) & 0x1f) << 8) + y_top),
+                            .z = static_cast<int16_t>((v & 0x1f) << 10)
+                         },
+                        .attributes = 0,
+                        .colour = trview::Colour(colour, colour, colour)
+                    };
+                }) | std::ranges::to<std::vector>();
+    }
+
     std::vector<trview_room_vertex> convert_vertices_tr3_psx(std::vector<uint32_t> vertices, int32_t y_top)
     {
         return vertices |
